+++ test/test_testdirectory.py
@@ -2,10 +2,13 @@ import os
 
 def test_run(testdirectory):
 
-    testdirectory.run('python','--version')
+    testdirectory.run(['python','--version'])
 
-    testdirectory.run('python','--version', stdout=None, stderr=None)
+    testdirectory.run(['python','--version'], stdout=None, stderr=None)
 
+    r = testdirectory.run('python --version')
+    assert r.returncode == 0
+    assert r.stdout.match('Python *') or r.stderr.match('Python *')
 
 def test_testdirectory(testdirectory):
     """ Unit test for the testdirectory fixture"""
@@ -34,11 +37,6 @@ def test_testdirectory(testdirectory):
     assert os.path.exists(os.path.join(sub2.path(), 'sub1'))
     assert os.path.exists(sub1_copy.path())
 
-    # Run a command that should be available on all platforms
-    r = sub1.run('python', '--version')
-
-    assert r.returncode == 0
-    assert r.stdout.match('Python *') or r.stderr.match('Python *')
 
     sub3 = testdirectory.mkdir('sub3')
     ok3_file = sub3.copy_file(ok_path, rename_as='ok3.txt')
@@ -65,3 +63,22 @@ def test_testdirectory(testdirectory):
     sub5.rmdir()
 
     assert not testdirectory.contains_dir(os.path.join('sub4', 'sub5'))
+
+def test_write_text(testdirectory):
+    ok_path = testdirectory.write_text('ok.txt', u'hello_world',
+                                       encoding='utf-8')
+
+    assert testdirectory.contains_file('ok.txt')
+    assert os.path.isfile(ok_path)
+
+def test_symlink(testdirectory):
+    sub1 = testdirectory.mkdir('sub1')
+    sub2 = testdirectory.mkdir('sub2')
+
+    ok_path = sub1.write_text('ok.txt', u'hello_world', encoding='utf-8')
+
+    # Create a symlink to 'ok.txt' inside sub2
+    link_path = sub2.symlink_file(ok_path)
+
+    assert sub2.contains_file('ok.txt')
+    assert os.path.isfile(link_path)